Changed crowd and thug scripts
Added the thug script and mostly cloned the fan script onto it for its normal behaviour (before collision with pla
yer). Also changed crowd generation to avoid fans merging. This last change is to ensure that when thugs are generated, they are free to move and chase after the player.